started working with cookies

+++ medshop/store/models.py
@@ -15,7 +15,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=150, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
     description = models.CharField(max_length=500, null=True)
     image = models.ImageField(null=True, blank=True)
